requirements.txt
git+https://github.com/gabrielusvicente/faker.git
arrow==0.7.0
Django==1.9.4
dj-database-url==0.4.0
django-compressor==2.0
django-extensions==1.6.1
django-ipware==1.1.3
django-test-without-migrations==0.4
gunicorn==19.4.5
libsass==0.10.1
psycopg2==2.6.1
python-decouple==3.0
python-social-auth==0.2.14
pytz
requests==2.9.1
whitenoise==2.0.6
